requirements.txt
tensorflow==2.6.0
keras==2.6.0
numpy==1.24.0
pandas==1.5.3
matplotlib==3.6.3
opencv-python==4.5.5.62
pillow==9.2.0
scipy==1.9.3
scikit-learn
deepface
